Revert "Change VirtualBox NIC"
This reverts commit 61813712d9fb68bf1baa9576723cae93bc79d004.

This change has been causing problems and unintended
consequences, such as affecting coreos NIC configuration
when running with parallels, even though this should have
only changed things with Virtualbox.  And while it
seems to have fixed the stability issues when running
with linux/virtualbox, it is not working the same
with mac/virtualbox.